split up unicode systems into different files
